Fix CardActivatorOnagerABasicTest
This changes the above file to correctly call the getActivator
method. The getActivator is equivalent to activating a
card. It should be called again for every activation, as per the
interface javadocs. It is not valid to call any methods on an
activator after the complete method has been called.